Main testing functionality now works.
Force the roothub that the device is under to be "on" and then enable
autosuspend for the device under test.  Cat the active_duration files
and for both, and calculate the delta time (with a little wiggle
room) that both devices have been active.  Use this to test when a
USB device is suspended or resumed.  Thanks to Oliver Neukum for the
suggestion!

Clean up the roothub's level files and the devices' level files after this test.

Signed-off-by: Sarah Sharp <sarah.a.sharp@intel.com>