Improved testing for active device.
Use the urbnum sysfs file as an indication of device activity, rather
than the active_duration file.  The number of URBs submitted to the
device will increase when the driver communicates with the device.

We use urbnum instead of active_duration so that we can sample the
activity before prompting the user to initiate activity on the device.
This allows the user to use the device for a short amount of time, and
then press enter.  Using active_duration would have been a pain, because
the user may take some time to find a program that uses their device and
calculating if a device was suspended for a very short amount of time
relative to the parent is a pain to figure out.

Signed-off-by: Sarah Sharp <sarah.a.sharp@intel.com>