Add a common exit routine.
Add a function that will be called whenever the script calls exit.  This
function will clean up any temporary files, and handle putting the USB sysfs
files back in order if something failed during the test.  The function will also
append the VID:PID to /etc/usb-pm-tool/pm-broken-devices.txt if the user reports
the device doesn't work, or the device disconnected during the test.  (I'm
assuming it disconnected because of the selective suspend command).

Signed-off-by: Sarah Sharp <sarah.a.sharp@intel.com>