Update where we write the udev rule.
The udev documentation in /etc/udev/rules.d/README says that the udev
daemon now uses inotify to track changes to the rules in
/etc/udev/rules.d/.  It will automatically load new rules when changes
to the files occur, but the files can't be a symlink.

Change the script to just write the file in /etc/udev/rules.d/, rather
than writing it in /etc/udev/ and symlinking it.  Also, stop restarting
udev, since it will detect the change in the rules file without a
restart.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>