Don't print lines that don't match.
If someone messes up the input file format, we don't want a udev rule to
be created that isn't parsable.  Use the -n 'no print' flag for sed, and
explicitly tell it to print the matched expressions with the 'p' flag.

We might still end up with a udev file with no rules at all, but that
should be harmless.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>